Fixes for tests, fix issue with django 1.7 registring user model, change revert_to.
So there is an issue with Python 2.6 and django-tablib that we should be aware, which is caused by dict comprehention.
Tox env for python 2.6 is returned. we can disable it after discusing.
Changes tox.ini to use appropriate command and fail explicitly instead of always showen us a green light.