Fix plugins forms (#95), respect app configs, filter not published/apphooked configs
This allows us to filter app configs correctly according to app config state and if it is app hooked check if page is published (for cases when aldryn-apphook-reload is not used on the project.
Also minor changes to templates to use get_title instead of unicode method, because then we can display handy app config information in forms, and still have control over title and translations.