Fix Django migrations by using strings instead of bytes for fields options
This is an issue for Python 3. Migrations created under Python 2 and
applied under Python 3 could have issues in some field types. It was
fixed in recent Django versions, but we have to adjust old migration
files manually.

The specific issue was with django CMS placeholders:
`"Placeholder identifier names need to be of type string"`.